refactor: update UserData import paths to use the correct domain entity
